add: attachment to message move: forum config upload folder to define.php

+++ admin/forum_config.php
@@ -2,29 +2,37 @@
 
 declare(strict_types = 1);
 
+use Pu239\Cache;
+use Pu239\Database;
+
 require_once INCL_DIR . 'function_html.php';
 require_once CLASS_DIR . 'class_check.php';
 $class = get_access(basename($_SERVER['REQUEST_URI']));
 class_check($class);
 $lang = array_merge($lang, load_language('ad_forum_config'));
-global $site_config;
+global $container, $site_config;
 
 $HTMLOUT = $time_drop_down = $accepted_file_extension = $accepted_file_types = $member_class_drop_down = '';
 $settings_saved = false;
-//=== be sure to set your id (below) in the DB. as well as setting your upload dir to something unique
 $config_id = 1;
-if (isset($_POST['do_it'])) {
-    $delete_for_real = isset($_POST['delete_for_real']) ? (int) $_POST['delete_for_real'] : 0;
-    $site_config['forum_config']['min_delete_view_class'] = isset($_POST['min_delete_view_class']) && valid_class((int) $_POST['min_delete_view_class']) ? (int) $_POST['min_delete_view_class'] : 0;
-    $readpost_expiry = isset($_POST['readpost_expiry']) ? (int) $_POST['readpost_expiry'] : 0;
-    $min_upload_class = isset($_POST['min_upload_class']) && valid_class((int) $_POST['min_upload_class']) ? (int) $_POST['min_upload_class'] : 0;
-    $accepted_file_extension = isset($_POST['accepted_file_extension']) ? preg_replace('/\s+/', '|', trim($_POST['accepted_file_extension'])) : '';
-    $accepted_file_types = isset($_POST['accepted_file_types']) ? preg_replace('/\s+/', '|', trim($_POST['accepted_file_types'])) : '';
-    $max_file_size = isset($_POST['max_file_size']) ? (int) $_POST['max_file_size'] : 0;
-    $upload_folder = isset($_POST['upload_folder']) ? htmlsafechars(trim($_POST['upload_folder'])) : '';
-    sql_query('UPDATE forum_config SET delete_for_real = ' . sqlesc($delete_for_real) . ', min_delete_view_class = ' . sqlesc($site_config['forum_config']['min_delete_view_class']) . ', readpost_expiry = ' . sqlesc($readpost_expiry) . ', min_upload_class = ' . sqlesc($min_upload_class) . ', accepted_file_extension = ' . sqlesc($accepted_file_extension) . ',  accepted_file_types = ' . sqlesc($accepted_file_types) . ', max_file_size = ' . $max_file_size . ', upload_folder = ' . sqlesc($upload_folder) . ' WHERE id=' . sqlesc($config_id));
+$fluent = $container->get(Database::class);
+$cache = $container->get(Cache::class);
+if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['do_it'])) {
+    $update = [
+        'delete_for_real' => isset($_POST['delete_for_real']) ? (int) $_POST['delete_for_real'] : 0,
+        'min_delete_view_class' => isset($_POST['min_delete_view_class']) && valid_class((int) $_POST['min_delete_view_class']) ? (int) $_POST['min_delete_view_class'] : 0,
+        'readpost_expiry' => isset($_POST['readpost_expiry']) ? (int) $_POST['readpost_expiry'] : 0,
+        'min_upload_class' => isset($_POST['min_upload_class']) && valid_class((int) $_POST['min_upload_class']) ? (int) $_POST['min_upload_class'] : 0,
+        'accepted_file_extension' => isset($_POST['accepted_file_extension']) ? preg_replace('/\s+/', '|', trim($_POST['accepted_file_extension'])) : '',
+        'accepted_file_types' => isset($_POST['accepted_file_types']) ? preg_replace('/\s+/', '|', trim($_POST['accepted_file_types'])) : '',
+        'max_file_size' => isset($_POST['max_file_size']) ? (int) $_POST['max_file_size'] : 0,
+    ];
+    $fluent->update('forum_config')
+           ->set($update)
+           ->where('id = ?', $config_id)
+           ->execute();
     $cache->delete('forum_config_');
-    header('Location: ' . $_SERVER['PHP_SELF'] . '?tool=forum_config&action=forum_config');
+    header('Location: ' . $_SERVER['PHP_SELF'] . '?tool=forum_config');
     die();
 }
 $main_links = "
@@ -41,7 +49,7 @@ $main_links = "
             <h1 class='has-text-centered'>{$lang['forum_config_config']}</h1>";
 
 $res = sql_query('SELECT delete_for_real, min_delete_view_class, readpost_expiry, min_upload_class, accepted_file_extension,
-                                accepted_file_types, max_file_size, upload_folder FROM forum_config WHERE id=' . sqlesc($config_id));
+                                accepted_file_types, max_file_size FROM forum_config WHERE id = ' . sqlesc($config_id));
 $arr = mysqli_fetch_array($res);
 $weeks = 1;
 for ($i = 7; $i <= 365; $i = $i + 7) {
@@ -50,7 +58,7 @@ for ($i = 7; $i <= 365; $i = $i + 7) {
 }
 $accepted_file_extension = (!empty($arr['accepted_file_extension'])) ? str_replace('|', ' ', $arr['accepted_file_extension']) : [];
 $accepted_file_types = (!empty($arr['accepted_file_types'])) ? str_replace('|', ' ', $arr['accepted_file_types']) : [];
-$HTMLOUT .= $main_links . '<form method="post" action="staffpanel.php?tool=forum_config&amp;action=forum_config" accept-charset="utf-8">
+$HTMLOUT .= $main_links . '<form method="post" action="staffpanel.php?tool=forum_config" accept-charset="utf-8">
             <input type="hidden" name="do_it" value="1">
         <table class="table table-bordered table-striped">
         <tr>
@@ -95,12 +103,6 @@ $HTMLOUT .= $main_links . '<form method="post" action="staffpanel.php?tool=forum
             <td>
             <input name="max_file_size" type="text" class="w-100" maxlength="200" value="' . (int) $arr['max_file_size'] . '"><br>
             ' . $lang['forum_config_size_desc'] . '' . mksize($arr['max_file_size']) . '.</td>
-         </tr>
-          <tr>
-            <td><span class="has-text-weight-bold">' . $lang['forum_config_folder'] . '</span>  </td>
-            <td>
-            <input name="upload_folder" type="text" class="w-100" maxlength="200" value="' . htmlsafechars($arr['upload_folder']) . '"><br>
-            ' . $lang['forum_config_folder_desc'] . '<br>
          </tr>
         <tr>
             <td colspan="2" class="has-text-centered">
